The board finds a shorted gate pair, and PE15 stopped disconnecting the break
W read 99.13 % both-gates-high against 0.00 % on U and V - its two pins were
one node, a bad footprint, since reworked. `Board_PwmGateShorts()` drives one
pin and watches the other sink through its own pull-down; `gate_shorts` is
appended to the gate driver state (proto MINOR 26) and `gate_drivers_check()`
refuses to arm on it. After the rework: 51.6 / 47.0 / 0.0 %, matching U and V,
and 20 s at 50 % on all three legs moved the NTC 29.52 -> 33.23 C.

PE15 carries TIM1_BKIN and was `usable = true`, so the test path called
HAL_GPIO_Init on it and took the break off the timer - measured after a
conformance run, MODER 00 with OTYPER and PUPDR still holding the AF_OD
setup. Now reserved, in both the board and the stand-in. Two test_mcp checks
moved to E14.

BKIN was also AF_OD with no pull, so an undriven fault line floated;
Board_PwmInit sets a pull-up. One -Wconversion warning in cmd_board.c.

+++ Comms/Src/cmd_board.c
@@ -68,7 +68,7 @@ static cmd_status_t h_adc_table(rd_t *in, wr_t *out)
   /* Counted first, then written, because the count leads the rows and a
      writer that ran out halfway would have already sent the wrong one. */
   uint8_t n = 0U;
-  uint16_t room = wr_room(out) - 2U;
+  uint16_t room = (uint16_t)(wr_room(out) - 2U);
 
   for (uint8_t i = start; i < total; i++)
   {
